Upload core code
CPU_ASM_parser.py is a parser for the assembly language used by the CPU. It is responsible for performing basic validation checks regarding the assembly and then converting it into a form that is convenient for execution.
CPU_core.py implements the CPU class which performs the actual execution of the assembly and logging of said execution. It also exposes some methods to dump out state and extract an append only log of the execution(both of which are reset upon loading a new assembly program).