Example assembly files
The files here are intended for example use and testing.
test_asm.txt is valid assembly intended to check that the parser is correctly processing valid input.
counter_asm.txt should run a counter and output the state of the counter into ascending memory addresses until the counter overflows. This can be used to verify that the code works.
check_carry_asm.txt is used to verify that the CARRY register is being corrected processed.